+++ src/include/LSP/ClientConfiguration.hpp
@@ -156,7 +156,7 @@ struct ClientCompletionImportsConfiguration
     /// Whether services and requires should be separated by an empty line
     bool separateGroupsWithLine = false;
     /// Files that match these globs will not be shown during auto-import
-    std::vector<std::string> ignoreGlobs{};
+    std::vector<std::string> ignoreGlobs{"**/_Index/**"};
 };
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(ClientCompletionImportsConfiguration, enabled, suggestServices, includedServices, excludedServices,
     suggestRequires, requireStyle, stringRequires, separateGroupsWithLine, ignoreGlobs);
@@ -274,7 +274,7 @@ struct ClientConfiguration
     /// Whether to automatically autocomplete end
     /// DEPRECATED: Use completion.autocompleteEnd instead
     bool autocompleteEnd = false;
-    std::vector<std::string> ignoreGlobs{};
+    std::vector<std::string> ignoreGlobs{"**/_Index/**"};
     ClientPlatformConfiguration platform{};
     ClientRobloxSourcemapConfiguration sourcemap{};
     ClientDiagnosticsConfiguration diagnostics{};