+++ tests/LuauExt.test.cpp
@@ -1,6 +1,7 @@
 #include "doctest.h"
 #include "Fixture.h"
 #include "LSP/LuauExt.hpp"
+#include "Platform/InstanceRequireAutoImporter.hpp"
 #include "Platform/RobloxPlatform.hpp"
 
 TEST_SUITE_BEGIN("Luau Extensions");
@@ -12,7 +13,7 @@ TEST_CASE_FIXTURE(Fixture, "FindImports service location 1")
     )");
     REQUIRE(block);
 
-    RobloxFindImportsVisitor visitor;
+    Luau::LanguageServer::AutoImports::RobloxFindImportsVisitor visitor;
     visitor.visit(block);
 
     CHECK_EQ(visitor.findBestLineForService("CollectionService", 0), 1);
@@ -27,7 +28,7 @@ TEST_CASE_FIXTURE(Fixture, "FindImports service location 2")
     )");
     REQUIRE(block);
 
-    RobloxFindImportsVisitor visitor;
+    Luau::LanguageServer::AutoImports::RobloxFindImportsVisitor visitor;
     visitor.visit(block);
 
     CHECK_EQ(visitor.findBestLineForService("CollectionService", 0), 1);
@@ -43,7 +44,7 @@ TEST_CASE_FIXTURE(Fixture, "FindImports service location 3")
     )");
     REQUIRE(block);
 
-    RobloxFindImportsVisitor visitor;
+    Luau::LanguageServer::AutoImports::RobloxFindImportsVisitor visitor;
     visitor.visit(block);
 
     CHECK_EQ(visitor.findBestLineForService("Workspace", 0), 3);